Add support for VMware Workstation to Fusion driver
