+++ drivers/vmwarefusion/fusion_test.go
@@ -1,3 +1,5 @@
+//+build !windows
+
 package vmwarefusion
 
 import (
