+++ drivers/vmwarefusion/fusion_unsupported.go
@@ -1,0 +1,9 @@
+// +build !darwin,!linux
+
+package vmwarefusion
+
+import "github.com/docker/machine/libmachine/drivers"
+
+func NewDriver(hostName, storePath string) drivers.Driver {
+	return drivers.NewDriverNotSupported("vmwarefusion", hostName, storePath)
+}
